Move seq benches to new file
